src/app/router/group.router.go
package router

import (
	"github.com/gofiber/fiber/v2"
	"github.com/isd-sgcu/rnkm65-gateway/src/app/handler/group"
)

func (r *FiberRouter) GetGroup(path string, h func(ctx group.IContext)) {
	r.group.Get(path, func(c *fiber.Ctx) error {
		h(NewFiberCtx(c))
		return nil
	})
}

func (r *FiberRouter) PostGroup(path string, h func(ctx group.IContext)) {
	r.group.Post(path, func(c *fiber.Ctx) error {
		h(NewFiberCtx(c))
		return nil
	})
}

func (r *FiberRouter) PutGroup(path string, h func(ctx group.IContext)) {
	r.group.Put(path, func(c *fiber.Ctx) error {
		h(NewFiberCtx(c))
		return nil
	})
}

func (r *FiberRouter) DeleteGroup(path string, h func(ctx group.IContext)) {
	r.group.Delete(path, func(c *fiber.Ctx) error {
		h(NewFiberCtx(c))
		return nil
	})
}
